Add per-box cascading collision group mask
If not set in the box def, the box inherits from the body.

Use case: a body with two attached boxes that have different
collision group requirements.

Example: a powerup that collides with the static world
but not with other dynamic objects -- so a player can run
over it without collision. A sensor is attached that will
collide with the player's box only, so you know when
you have picked it up.